move MSW keyword handlers to separate file
